%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% The String class
\alias{String}
\alias{as.character.String}
\alias{length.String}
\alias{print.String}
\alias{charAt}
\alias{charAt.default}
\alias{compareTo}
\alias{compareTo.default}
\alias{concat}
\alias{concat.default}
\alias{endsWith}
\alias{endsWith.default}
\alias{equalsIgnoreCase}
\alias{equalsIgnoreCase.default}
\alias{indexOf}
\alias{indexOf.default}
\alias{lastIndexOf}
\alias{lastIndexOf.default}
\alias{nchar}
\alias{nchar.default}
\alias{regionMatches}
\alias{regionMatches.default}
\alias{replace}
\alias{replace.default}
\alias{set}
\alias{startsWith}
\alias{startsWith.default}
\alias{substring}
\alias{substring.default}
\alias{toCharArray}
\alias{toCharArray.default}
\alias{toLowerCase}
\alias{toLowerCase.default}
\alias{toUpperCase}
\alias{toUpperCase.default}
\alias{valueOf}
\alias{getBytes}
\alias{getBytes.default}

% Functions
\alias{as.Java.byte}
\alias{as.Java.byte.default}
\alias{as.Java.int}
\alias{as.Java.int.default}
\alias{as.Java.long}
\alias{as.Java.long.default}
\alias{as.Java.short}
\alias{as.Java.short.default}
\alias{getBit}
\alias{getBit.default}
\alias{getBits}
\alias{getBits.default}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
